refactor(aztec-node): extract world-state and witness queries into a module
Move the node's Merkle-tree and membership-witness queries (findLeavesIndexes, the note-hash /
block-hash / L1-to-L2 / nullifier / low-nullifier witnesses, getPublicDataWitness,
getPublicStorageAt, getL2ToL1Messages, getL2ToL1MembershipWitness, getL1ToL2MessageCheckpoint)
together with the block-resolution and reorg-aware sync engine (getWorldState, resolveBlockNumber,
syncWorldState) into a new `NodeWorldStateQueries` collaborator under `src/modules/`.

`AztecNodeService` keeps thin delegating methods (mirroring `nodePublicCallsSimulator`), so the public
and RPC surface is unchanged. `getWorldState` remains as a protected delegate so subclasses and tests can
still exercise the node's sync behavior. `normalizeBlockParameter` is injected into the module for now; a
follow-up extracts it as a free function.
