+++ yarn-project/aztec-node/src/aztec-node/server.ts
@@ -13,7 +13,7 @@ import {
   EpochNumber,
   SlotNumber,
 } from '@aztec/foundation/branded-types';
-import { chunkBy, compactArray, pick, unique } from '@aztec/foundation/collection';
+import { compactArray, pick, unique } from '@aztec/foundation/collection';
 import { Fr } from '@aztec/foundation/curves/bn254';
 import { EthAddress } from '@aztec/foundation/eth-address';
 import { BadRequestError } from '@aztec/foundation/json-rpc';
@@ -44,7 +44,6 @@ import {
   type L2BlockTag,
   type L2Tips,
   type NormalizedBlockParameter,
-  inspectBlockParameter,
 } from '@aztec/stdlib/block';
 import { type CheckpointData, L1PublishedData, type PublishedCheckpoint } from '@aztec/stdlib/checkpoint';
 import type {
@@ -56,7 +55,6 @@ import type {
 } from '@aztec/stdlib/contract';
 import { getEpochAtSlot } from '@aztec/stdlib/epoch-helpers';
 import { GasFees, type ManaUsageEstimate, getNetworkTxGasLimits } from '@aztec/stdlib/gas';
-import { computePublicDataTreeLeafSlot } from '@aztec/stdlib/hash';
 import type {
   AztecNode,
   AztecNodeAdmin,
@@ -85,10 +83,9 @@ import {
 } from '@aztec/stdlib/interfaces/server';
 import type { DebugLogStore, LogResult, PrivateLogsQuery, PublicLogsQuery } from '@aztec/stdlib/logs';
 import { NullDebugLogStore } from '@aztec/stdlib/logs';
-import { InboxLeaf, type L1ToL2MessageSource, type L2ToL1MembershipWitness } from '@aztec/stdlib/messaging';
+import type { L1ToL2MessageSource, L2ToL1MembershipWitness } from '@aztec/stdlib/messaging';
 import type { CheckpointAttestation } from '@aztec/stdlib/p2p';
 import type { Offense } from '@aztec/stdlib/slashing';
-import type { NullifierLeafPreimage, PublicDataTreeLeafPreimage } from '@aztec/stdlib/trees';
 import { MerkleTreeId, NullifierMembershipWitness, PublicDataWitness } from '@aztec/stdlib/trees';
 import {
   DroppedTxReceipt,
@@ -118,6 +115,7 @@ import {
 } from '@aztec/telemetry-client';
 import { NodeKeystoreAdapter, ValidatorClient } from '@aztec/validator-client';
 
+import { NodeWorldStateQueries } from '../modules/node_world_state_queries.js';
 import { Sentinel } from '../sentinel/sentinel.js';
 import {
   blockResponseFromBlockData,
@@ -176,6 +174,7 @@ export class AztecNodeService implements AztecNode, AztecNodeAdmin, AztecNodeDeb
   // Saved minTxsPerBlock used by `pauseSequencer` to restore production-sequencer config on resume.
   private sequencerPausedMinTxsPerBlock: number | undefined;
   private readonly nodePublicCallsSimulator: NodePublicCallsSimulator;
+  private readonly worldStateQueries: NodeWorldStateQueries;
 
   public readonly tracer: Tracer;
 
@@ -255,6 +254,14 @@ export class AztecNodeService implements AztecNode, AztecNodeAdmin, AztecNodeDeb
       config: this.config,
       telemetry: this.telemetry,
       log: this.log.createChild('public-calls-simulator'),
+    });
+
+    this.worldStateQueries = new NodeWorldStateQueries({
+      worldStateSynchronizer: this.worldStateSynchronizer,
+      blockSource: this.blockSource,
+      l1ToL2MessageSource: this.l1ToL2MessageSource,
+      normalizeBlockParameter: param => this.normalizeBlockParameter(param),
+      log: this.log.createChild('world-state-queries'),
     });
 
     this.log.info(`Aztec Node version: ${this.packageVersion}`);
@@ -855,125 +862,37 @@ export class AztecNodeService implements AztecNode, AztecNodeAdmin, AztecNodeDeb
     return compactArray(txs);
   }
 
-  public async findLeavesIndexes(
+  public findLeavesIndexes(
     referenceBlock: BlockParameter,
     treeId: MerkleTreeId,
     leafValues: Fr[],
   ): Promise<(DataInBlock<bigint> | undefined)[]> {
-    const committedDb = await this.getWorldState(referenceBlock);
-    const maybeIndices = await committedDb.findLeafIndices(
-      treeId,
-      leafValues.map(x => x.toBuffer()),
-    );
-    // Filter out undefined values to query block numbers only for found leaves
-    const definedIndices = maybeIndices.filter(x => x !== undefined);
-
-    // Now we find the block numbers for the defined indices
-    const blockNumbers = await committedDb.getBlockNumbersForLeafIndices(treeId, definedIndices);
-
-    // Build a map from leaf index to block number
-    const indexToBlockNumber = new Map<bigint, BlockNumber>();
-    for (let i = 0; i < definedIndices.length; i++) {
-      const blockNumber = blockNumbers[i];
-      if (blockNumber === undefined) {
-        throw new Error(
-          `Block number is undefined for leaf index ${definedIndices[i]} in tree ${MerkleTreeId[treeId]}`,
-        );
-      }
-      indexToBlockNumber.set(definedIndices[i], blockNumber);
-    }
-
-    // Get unique block numbers in order to optimize num calls to getLeafValue function.
-    const uniqueBlockNumbers = [...new Set(indexToBlockNumber.values())];
-
-    // Now we obtain the block hashes from the archive tree (block number = leaf index in archive tree).
-    const blockHashes = await Promise.all(
-      uniqueBlockNumbers.map(blockNumber => {
-        return committedDb.getLeafValue(MerkleTreeId.ARCHIVE, BigInt(blockNumber));
-      }),
-    );
-
-    // Build a map from block number to block hash
-    const blockNumberToHash = new Map<BlockNumber, BlockHash>();
-    for (let i = 0; i < uniqueBlockNumbers.length; i++) {
-      const blockHash = blockHashes[i];
-      if (blockHash === undefined) {
-        throw new Error(`Block hash is undefined for block number ${uniqueBlockNumbers[i]}`);
-      }
-      blockNumberToHash.set(uniqueBlockNumbers[i], blockHash);
-    }
-
-    // Create DataInBlock objects by combining indices, blockNumbers and blockHashes and return them.
-    return maybeIndices.map(index => {
-      if (index === undefined) {
-        return undefined;
-      }
-      const blockNumber = indexToBlockNumber.get(index);
-      if (blockNumber === undefined) {
-        throw new Error(`Block number not found for leaf index ${index} in tree ${MerkleTreeId[treeId]}`);
-      }
-      const l2BlockHash = blockNumberToHash.get(blockNumber);
-      if (l2BlockHash === undefined) {
-        throw new Error(`Block hash not found for block number ${blockNumber}`);
-      }
-      return {
-        l2BlockNumber: blockNumber,
-        l2BlockHash,
-        data: index,
-      };
-    });
+    return this.worldStateQueries.findLeavesIndexes(referenceBlock, treeId, leafValues);
   }
 
-  public async getBlockHashMembershipWitness(
+  public getBlockHashMembershipWitness(
     referenceBlock: BlockParameter,
     blockHash: BlockHash,
   ): Promise<MembershipWitness<typeof ARCHIVE_HEIGHT> | undefined> {
-    // The Noir circuit checks the archive membership proof against `anchor_block_header.last_archive.root`,
-    // which is the archive tree root BEFORE the anchor block was added (i.e. the state after block N-1).
-    // So we need the world state at block N-1, not block N, to produce a sibling path matching that root.
-    const referenceBlockNumber = await this.resolveBlockNumber(referenceBlock);
-    if (referenceBlockNumber === BlockNumber.ZERO) {
-      // Block 0 (the initial block) has an empty archive, so no membership witness can exist.
-      return undefined;
-    }
-    const committedDb = await this.getWorldState(BlockNumber(referenceBlockNumber - 1));
-    const [pathAndIndex] = await committedDb.findSiblingPaths<MerkleTreeId.ARCHIVE>(MerkleTreeId.ARCHIVE, [blockHash]);
-    return pathAndIndex === undefined
-      ? undefined
-      : MembershipWitness.fromSiblingPath(pathAndIndex.index, pathAndIndex.path);
+    return this.worldStateQueries.getBlockHashMembershipWitness(referenceBlock, blockHash);
   }
 
-  public async getNoteHashMembershipWitness(
+  public getNoteHashMembershipWitness(
     referenceBlock: BlockParameter,
     noteHash: Fr,
   ): Promise<MembershipWitness<typeof NOTE_HASH_TREE_HEIGHT> | undefined> {
-    const committedDb = await this.getWorldState(referenceBlock);
-    const [pathAndIndex] = await committedDb.findSiblingPaths<MerkleTreeId.NOTE_HASH_TREE>(
-      MerkleTreeId.NOTE_HASH_TREE,
-      [noteHash],
-    );
-    return pathAndIndex === undefined
-      ? undefined
-      : MembershipWitness.fromSiblingPath(pathAndIndex.index, pathAndIndex.path);
+    return this.worldStateQueries.getNoteHashMembershipWitness(referenceBlock, noteHash);
   }
 
-  public async getL1ToL2MessageMembershipWitness(
+  public getL1ToL2MessageMembershipWitness(
     referenceBlock: BlockParameter,
     l1ToL2Message: Fr,
   ): Promise<[bigint, SiblingPath<typeof L1_TO_L2_MSG_TREE_HEIGHT>] | undefined> {
-    const db = await this.getWorldState(referenceBlock);
-    const [witness] = await db.findSiblingPaths(MerkleTreeId.L1_TO_L2_MESSAGE_TREE, [l1ToL2Message]);
-    if (!witness) {
-      return undefined;
-    }
-
-    // REFACTOR: Return a MembershipWitness object
-    return [witness.index, witness.path];
+    return this.worldStateQueries.getL1ToL2MessageMembershipWitness(referenceBlock, l1ToL2Message);
   }
 
-  public async getL1ToL2MessageCheckpoint(l1ToL2Message: Fr): Promise<CheckpointNumber | undefined> {
-    const messageIndex = await this.l1ToL2MessageSource.getL1ToL2MessageIndex(l1ToL2Message);
-    return messageIndex !== undefined ? InboxLeaf.checkpointNumberFromIndex(messageIndex) : undefined;
+  public getL1ToL2MessageCheckpoint(l1ToL2Message: Fr): Promise<CheckpointNumber | undefined> {
+    return this.worldStateQueries.getL1ToL2MessageCheckpoint(l1ToL2Message);
   }
 
   /**
@@ -984,12 +903,8 @@ export class AztecNodeService implements AztecNode, AztecNodeAdmin, AztecNodeDeb
    * @param epoch - The epoch at which to get the data.
    * @returns The L2 to L1 messages (empty array if the epoch is not found).
    */
-  public async getL2ToL1Messages(epoch: EpochNumber): Promise<Fr[][][][]> {
-    const blocks = await this.blockSource.getBlocks({ epoch, onlyCheckpointed: true });
-    const blocksInCheckpoints = chunkBy(blocks, block => block.header.globalVariables.slotNumber);
-    return blocksInCheckpoints.map(slotBlocks =>
-      slotBlocks.map(block => block.body.txEffects.map(txEffect => txEffect.l2ToL1Msgs)),
-    );
+  public getL2ToL1Messages(epoch: EpochNumber): Promise<Fr[][][][]> {
+    return this.worldStateQueries.getL2ToL1Messages(epoch);
   }
 
   /**
@@ -1001,77 +916,29 @@ export class AztecNodeService implements AztecNode, AztecNodeAdmin, AztecNodeDeb
     message: Fr,
     messageIndexInTx?: number,
   ): Promise<L2ToL1MembershipWitness | undefined> {
-    return this.blockSource.getL2ToL1MembershipWitness(txHash, message, messageIndexInTx);
+    return this.worldStateQueries.getL2ToL1MembershipWitness(txHash, message, messageIndexInTx);
   }
 
-  public async getNullifierMembershipWitness(
+  public getNullifierMembershipWitness(
     referenceBlock: BlockParameter,
     nullifier: Fr,
   ): Promise<NullifierMembershipWitness | undefined> {
-    const db = await this.getWorldState(referenceBlock);
-    const [witness] = await db.findSiblingPaths(MerkleTreeId.NULLIFIER_TREE, [nullifier.toBuffer()]);
-    if (!witness) {
-      return undefined;
-    }
-
-    const { index, path } = witness;
-    const leafPreimage = await db.getLeafPreimage(MerkleTreeId.NULLIFIER_TREE, index);
-    if (!leafPreimage) {
-      return undefined;
-    }
-
-    return new NullifierMembershipWitness(index, leafPreimage as NullifierLeafPreimage, path);
+    return this.worldStateQueries.getNullifierMembershipWitness(referenceBlock, nullifier);
   }
 
-  public async getLowNullifierMembershipWitness(
+  public getLowNullifierMembershipWitness(
     referenceBlock: BlockParameter,
     nullifier: Fr,
   ): Promise<NullifierMembershipWitness | undefined> {
-    const committedDb = await this.getWorldState(referenceBlock);
-    const findResult = await committedDb.getPreviousValueIndex(MerkleTreeId.NULLIFIER_TREE, nullifier.toBigInt());
-    if (!findResult) {
-      return undefined;
-    }
-    const { index, alreadyPresent } = findResult;
-    if (alreadyPresent) {
-      throw new Error(
-        `Cannot prove nullifier non-inclusion: nullifier ${nullifier.toBigInt()} already exists in the tree`,
-      );
-    }
-    const preimageData = (await committedDb.getLeafPreimage(MerkleTreeId.NULLIFIER_TREE, index))!;
-
-    const siblingPath = await committedDb.getSiblingPath(MerkleTreeId.NULLIFIER_TREE, BigInt(index));
-    return new NullifierMembershipWitness(BigInt(index), preimageData as NullifierLeafPreimage, siblingPath);
+    return this.worldStateQueries.getLowNullifierMembershipWitness(referenceBlock, nullifier);
   }
 
-  async getPublicDataWitness(referenceBlock: BlockParameter, leafSlot: Fr): Promise<PublicDataWitness | undefined> {
-    const committedDb = await this.getWorldState(referenceBlock);
-    const lowLeafResult = await committedDb.getPreviousValueIndex(MerkleTreeId.PUBLIC_DATA_TREE, leafSlot.toBigInt());
-    if (!lowLeafResult) {
-      return undefined;
-    } else {
-      const preimage = (await committedDb.getLeafPreimage(
-        MerkleTreeId.PUBLIC_DATA_TREE,
-        lowLeafResult.index,
-      )) as PublicDataTreeLeafPreimage;
-      const path = await committedDb.getSiblingPath(MerkleTreeId.PUBLIC_DATA_TREE, lowLeafResult.index);
-      return new PublicDataWitness(lowLeafResult.index, preimage, path);
-    }
+  public getPublicDataWitness(referenceBlock: BlockParameter, leafSlot: Fr): Promise<PublicDataWitness | undefined> {
+    return this.worldStateQueries.getPublicDataWitness(referenceBlock, leafSlot);
   }
 
-  public async getPublicStorageAt(referenceBlock: BlockParameter, contract: AztecAddress, slot: Fr): Promise<Fr> {
-    const committedDb = await this.getWorldState(referenceBlock);
-    const leafSlot = await computePublicDataTreeLeafSlot(contract, slot);
-
-    const lowLeafResult = await committedDb.getPreviousValueIndex(MerkleTreeId.PUBLIC_DATA_TREE, leafSlot.toBigInt());
-    if (!lowLeafResult || !lowLeafResult.alreadyPresent) {
-      return Fr.ZERO;
-    }
-    const preimage = (await committedDb.getLeafPreimage(
-      MerkleTreeId.PUBLIC_DATA_TREE,
-      lowLeafResult.index,
-    )) as PublicDataTreeLeafPreimage;
-    return preimage.leaf.value;
+  public getPublicStorageAt(referenceBlock: BlockParameter, contract: AztecAddress, slot: Fr): Promise<Fr> {
+    return this.worldStateQueries.getPublicStorageAt(referenceBlock, contract, slot);
   }
 
   /**
@@ -1475,91 +1342,11 @@ export class AztecNodeService implements AztecNode, AztecNodeAdmin, AztecNodeDeb
   }
 
   /**
-   * Returns an instance of MerkleTreeOperations having first ensured the world state is fully synched
-   * @param block - The block parameter (block number, block hash, or 'latest') at which to get the data.
-   * @returns An instance of a committed MerkleTreeOperations
+   * Returns a committed world-state view at `block`, driving sync first. Delegates to
+   * {@link NodeWorldStateQueries.getWorldState}; kept as a protected method so subclasses and tests can
+   * exercise the node's block-resolution and reorg-aware sync behavior.
    */
-  protected async getWorldState(block: BlockParameter) {
-    const query = this.normalizeBlockParameter(block);
-
-    // When the request anchors on a specific block hash, resolve it against the archiver up front and
-    // drive the world-state sync to that exact block number and hash. Resolving against the archiver
-    // first fails fast with a clear reorg error if the hash is unknown, and passing the hash to the
-    // synchronizer makes the sync reorg-aware: it barriers until the archive-tree commit for that block
-    // has landed and verifies it matches the requested fork, instead of syncing to bare latest height
-    // and then racing the snapshot read below against an in-flight archive-tree write.
-    const requestedHash = 'hash' in query ? query.hash : undefined;
-    const anchorBlockNumber = requestedHash !== undefined ? await this.resolveBlockNumber(query) : undefined;
-
-    let blockSyncedTo: BlockNumber = BlockNumber.ZERO;
-    try {
-      // Attempt to sync the world state if necessary
-      blockSyncedTo = await this.#syncWorldState(anchorBlockNumber, requestedHash);
-    } catch (err) {
-      this.log.error(`Error getting world state: ${err}`);
-    }
-
-    if ('tag' in query && query.tag === 'proposed') {
-      this.log.debug(`Using committed db for latest block, world state synced upto ${blockSyncedTo}`);
-      return this.worldStateSynchronizer.getCommitted();
-    }
-
-    const blockNumber = anchorBlockNumber ?? (await this.resolveBlockNumber(query));
-
-    // Check it's within world state sync range
-    if (blockNumber > blockSyncedTo) {
-      throw new Error(
-        `Queried block ${inspectBlockParameter(block)} not yet synced by the node (node is synced upto ${blockSyncedTo}).`,
-      );
-    }
-    this.log.debug(`Using snapshot for block ${blockNumber}, world state synced upto ${blockSyncedTo}`);
-
-    const snapshot = this.worldStateSynchronizer.getSnapshot(blockNumber);
-
-    // Double-check world-state synced to the same block hash as was requested.
-    // Block 0 is skipped: the snapshot returned by `getSnapshot(0)` is the *pre*-genesis archive
-    // (size 0), so leaf 0 is not yet inserted from that snapshot's view even though block 0's hash
-    // does live at archive index 0 in the committed tree. The genesis hash is already validated by
-    // the archiver when it resolves the hash query to block number 0.
-    if (requestedHash !== undefined && blockNumber !== BlockNumber.ZERO) {
-      const blockHash = await snapshot.getLeafValue(MerkleTreeId.ARCHIVE, BigInt(blockNumber));
-      if (!blockHash || !requestedHash.equals(blockHash)) {
-        throw new Error(
-          `Block hash ${requestedHash.toString()} not found in world state at block number ${blockNumber} (world state has ${blockHash?.toString() ?? 'no hash'} at that index, genesis header hash is ${this.blockSource.getGenesisBlockHash().toString()}). If the node API has been queried with anchor block hash possibly a reorg has occurred.`,
-        );
-      }
-    }
-
-    return snapshot;
-  }
-
-  /** Resolves any {@link BlockParameter} variant to a concrete block number. */
-  protected async resolveBlockNumber(block: BlockParameter): Promise<BlockNumber> {
-    const query = this.normalizeBlockParameter(block);
-    const blockNumber = await this.blockSource.getBlockNumber(query);
-    if (blockNumber === undefined) {
-      if ('hash' in query) {
-        throw new Error(
-          `Block hash ${query.hash.toString()} not found when querying world state. If the node API has been queried with anchor block hash possibly a reorg has occurred.`,
-        );
-      }
-      if ('archive' in query) {
-        throw new Error(`Block with archive ${query.archive.toString()} not found.`);
-      }
-      throw new Error(`Block not found for ${inspectBlockParameter(block)}.`);
-    }
-    return blockNumber;
-  }
-
-  /**
-   * Ensure the world state is synced.
-   * @param targetBlockNumber - Block to sync up to. Defaults to the latest block known to the archiver.
-   * @param blockHash - If provided, the synchronizer verifies the block at `targetBlockNumber` matches this
-   * hash, resyncing (and so detecting reorgs) if it does not yet match or has been reorged away.
-   * @returns A promise that fulfils once the world state is synced
-   */
-  async #syncWorldState(targetBlockNumber?: BlockNumber, blockHash?: BlockHash): Promise<BlockNumber> {
-    const target = targetBlockNumber ?? (await this.blockSource.getBlockNumber());
-    return await this.worldStateSynchronizer.syncImmediate(target, blockHash);
+  protected getWorldState(block: BlockParameter) {
+    return this.worldStateQueries.getWorldState(block);
   }
 }
